Through-The-Looking-Glass/python-app/requirements.txt
flask
python-logging-loki
prometheus-client
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp
